Asset Operations collections code clean-up
