+++ novawallet/Modules/AssetsSearch/AssetOperation/AssetOperationViewController.swift
@@ -1,0 +1,10 @@
+class AssetOperationViewController: AssetsSearchViewController {
+    override func setupCollectionManager() {
+        collectionViewManager = AssetOperationCollectionManager(
+            view: rootView,
+            groupsViewModel: groupsViewModel,
+            delegate: self,
+            selectedLocale: selectedLocale
+        )
+    }
+}
