+++ novawallet/Modules/AssetsSearch/AssetOperation/Send/SendAssetOperationCollectionManager/SendAssetOperationCollectionManager.swift
@@ -1,6 +1,6 @@
 import UIKit
 
-class SendAssetOperationCollectionManager: AssetsSearchCollectionManager {
+class SendAssetOperationCollectionManager: AssetOperationCollectionManager {
     var dataSource: SendAssetOperationCollectionDataSource? {
         get {
             collectionViewDataSource as? SendAssetOperationCollectionDataSource
@@ -37,42 +37,14 @@ class SendAssetOperationCollectionManager: AssetsSearchCollectionManager {
     }
 
     override func setup() {
-//        let layout = SendAssetOperationTokensFlowLayout()
-//        view?.collectionTokenGroupsLayout = layout
-//        view?.collectionView.setCollectionViewLayout(layout, animated: false)
-
         dataSource?.groupsLayoutDelegate = self
         dataSource?.delegate = self
 
         collectionViewDelegate.selectionDelegate = self
         collectionViewDelegate.groupsLayoutDelegate = self
 
-        view?.collectionView.dataSource = collectionViewDataSource
+        view?.collectionView.dataSource = dataSource
         view?.collectionView.delegate = collectionViewDelegate
-    }
-
-    override func updateTokensGroupLayout() {
-        guard
-            let tokenGroupsLayout,
-            groupsViewModel.listGroupStyle == .tokens
-        else {
-            return
-        }
-
-        groupsViewModel.listState.groups.enumerated().forEach { _, group in
-            guard case let .token(groupViewModel) = group else {
-                return
-            }
-
-            tokenGroupsLayout.setExpandableSection(
-                for: groupViewModel.token.symbol,
-                false
-            )
-        }
-    }
-
-    override func groupExpandable(for _: String) -> Bool {
-        false
     }
 }
